nice

+++ bot.js
@@ -22,18 +22,18 @@ const PREFIX = process.env.PREFIX || "!";
 const client = new Discord.Client();
 
 // Embed Presets
-const PING_EMBED = new Discord.RichEmbed().setTitle("Ping Results").setColor(0x21f8ff).addField("Latency", 0).addField("Discord API Latency", 0);
+const PING_EMBED = new Discord.RichEmbed().setTitle("Ping Results").setColor(0x21f8ff).addField("Latency", 0).addField("API Latency", 0);
 const HELP_EMBED = new Discord.RichEmbed().setTitle("Help").setColor(0x21f8ff)
-    .addField(`${PREFIX}ping`, "Get the bot's ping")
-    .addField(`${PREFIX}music [play|skip|import|stop|pause]`, "Didn't see it coming, did you? I am a music bot")
-    .addField(`${PREFIX}covid [country name]`, "Get COVID-19 data of a country")
+    .addField(`${PREFIX}ping`, "Get the bot's ping", true)
+    .addField(`${PREFIX}music [play|skip|import|stop|pause]`, "I am a music bot", true)
+    .addField(`${PREFIX}covid [country name]`, "Get COVID-19 data", true)
     .addField(`${PREFIX}azurlane`, `Find information on your ship waifu\nUsage: \`${PREFIX}azurlane (ship-name)\``)
     .addField(`${PREFIX}azurlane find`, `Query for a list of ships\nUsage: \`${PREFIX}azurlane find (field=value)/(field>value)+\`\nSeperated by ',' (comma)`)
     // .addField(`${PREFIX}coursemology`, `Access Coursemology.\nUsage: \`${PREFIX}coursemology (info|list|user) [args]\``)
     // .addField(`${PREFIX}timetable`, `Access time table\nUsage: \`${PREFIX}timetable [next|now]? [classname]?\``)
     .addField(`${PREFIX}24`, `Play 24 the number game\nUsage: \`${PREFIX}24\`, \`${PREFIX}24 impossible\``)
-    .addField(`${PREFIX}minesweeper`, `Play minesweeper the mine game\nUsage: \`${PREFIX}minesweeper <size>\`, \`<x> <y>\``)
-    .addField(`${PREFIX}gomoku`, `Play gomoku the monotone game\nUsage: \`${PREFIX}gomoku start <size> <win>\`, \`<x> <y>\``)
+    .addField(`${PREFIX}minesweeper`, `Play minesweeper the mine game\nUsage: \`${PREFIX}minesweeper size\`, \`x y\``)
+    .addField(`${PREFIX}gomoku`, `Play gomoku the monotone game\nUsage: \`${PREFIX}gomoku start size goal\`, \`x y\``)
     .addField(`${PREFIX}sleep`, "Tell you whether or not you should sleep.")
 chainUpStdOut();
 console.log('====== ZY Discord Bot Started! ======');
@@ -42,12 +42,12 @@ console.log('====== ZY Discord Bot Started! ======');
 client.on('ready', () => {
     console.log("=> Bot Running in " + client.guilds.keyArray().length + " servers!");
     client.guilds.get('665471208757657620').channels.get('665471209277882400').send("READY!");
-    client.user.setPresence(config.PRESENCE);
+    client.user.setPresence(config.PRESENCE).then(r => console.log("Presence Set!"));
     config.offset = 8 + new Date().getTimezoneOffset() / 60;
     config.HOOK = new Discord.WebhookClient('644427303719403521', process.env.HKTOKEN);
     config.HOOK2 = new Discord.WebhookClient('676309488021798912', process.env.HKTOKEN2);
     config.id = client.user.id;
-    azurlane.init();
+    azurlane.init().then(r => console.log("Azurlane Init!"));
     // coursemology.init();
     // setInterval(() => coursemology.update(config.DEFAULT_COURSE), 20000);
     const covidChannel = client.guilds.get('642273802520231936').channels.get('693051246885470209');
@@ -57,12 +57,12 @@ client.on('ready', () => {
     }
 });
 
-const OWO_24_MESSAGE_REGEX = /\*\*Ok <@!?\d+>, your numbers are: \`(\d \d \d \d)\` \*\*/g;
+const OWO_24_MESSAGE_REGEX = /\*\*Ok <@!?\d+>, your numbers are: `(\d \d \d \d)` \*\*/g;
 
 client.on('message', async msg => {
     let startTime = Date.now();
     console.log(`=> Message "${msg.content.replace('\n', '\\n').substring(0, 80) + (msg.content.length > 80 ? "..." : "")}" received from ${msg.author.tag}.`);
-    if (!msg.channel.type === "text") return;
+    if (msg.channel.type !== "text") return;
     if (!msg.guild && msg.author.id !== "456001047756800000") return; // Direct message from myself only
     if (msg.author.id === config.id) return; // Dont respond to bot's own messages
     if (msg.channel.id === "720891248352952341" && !process.env.LOCAL) return; // skip if not testing locally
@@ -87,11 +87,11 @@ client.on('message', async msg => {
     let args = msg.content.slice(1).trim().split(/\s+/g);
     const command = args.shift().toLowerCase();
     console.log(`running "${command}", args = [${args.join(", ")}]...`);
-    if (command === "help") msg.channel.send(HELP_EMBED);
+    if (command === "help") await msg.channel.send(HELP_EMBED);
     if (command === "ping") {
         const m = await msg.channel.send("Ping?");
-        PING_EMBED.fields[0].value = m.createdTimestamp - msg.createdTimestamp;
-        PING_EMBED.fields[1].value = Math.round(client.ping);
+        PING_EMBED.fields[0].value = `${m.createdTimestamp - msg.createdTimestamp}ms`;
+        PING_EMBED.fields[1].value = `${Math.round(client.ping)}ms`;
         PING_EMBED.setFooter("Requested By " + msg.author.username, msg.author.displayAvatarURL);
         console.log(` ping results obtained. lat = ${m.createdTimestamp - msg.createdTimestamp}, discord lat = ${Math.round(client.ping)}`);
         await m.edit(PING_EMBED);
@@ -99,7 +99,7 @@ client.on('message', async msg => {
 
     // if (command === "coursemology" || command === "cm") coursemology.handleCommand(args, msg);
     if (command === "list-emotes" || command === "emotes") {
-        sendLongMessage(msg.channel, "**Emotes:**\n" + join(msg.guild.emojis.array().map(e => `<:${e.name}:${e.id}> \`:${e.name}:\``), [" ", " ", "\n"]));
+        await sendLongMessage(msg.channel, "**Emotes:**\n" + join(msg.guild.emojis.array().map(e => `<:${e.name}:${e.id}> \`:${e.name}:\``), [" ", " ", "\n"]));
     }
     if (command === "pop" || command === "bubble" || command === "bubbles" || command === "bubble-pop") {
         sendBubblePop(msg, args);
@@ -107,9 +107,9 @@ client.on('message', async msg => {
     if (command === "anime" || command === "a") await anime.handleCommand(args, msg, PREFIX);
     if (command === "ms" || command === "minesweeper") minesweeper.handleCommand(args, msg, PREFIX);
     if (command === "minecraft") minecraft.handleCommand(args, msg, PREFIX);
-    if (command === "covid" || command === "coronavirus" || command === "corona" || command === "c") covid.handleCommand(args, msg, PREFIX);
-    if (command === "inspire") inspire.handleCommand(args, msg, PREFIX);
-    if (command === "azurlane" || command === "al" || command === "azur" || command === "az") azurlane.handleCommand(args, msg, PREFIX);
+    if (command === "covid" || command === "coronavirus" || command === "corona" || command === "c") await covid.handleCommand(args, msg, PREFIX);
+    if (command === "inspire") await inspire.handleCommand(args, msg, PREFIX);
+    if (command === "azurlane" || command === "al" || command === "azur" || command === "az") await azurlane.handleCommand(args, msg, PREFIX);
     if (command === "music" || command === "am" || command === "m" || command === "song") await music.handleCommand(args, msg, PREFIX);
     if (command.startsWith("!")) {
         console.log("Double !!: end command = " + command.substring(1));
@@ -126,7 +126,7 @@ client.on('message', async msg => {
     if (msg.author.id === "456001047756800000" && (command === "toggledebug" || command === "td")) {
         config.debug = !config.debug;
         console.log("DEBUG TOGGLED, debug = " + config.debug)
-        msg.channel.send(`DEBUG: debug output has been turned ${config.debug ? "on" : "off"}!`);
+        await msg.channel.send(`DEBUG: debug output has been turned ${config.debug ? "on" : "off"}!`);
     }
     if (msg.author.id === "456001047756800000" && (command === "forcestop" || command === "fs" || command === "restart" || command === "rs")) {
         console.log("Restarting program due to request from owner...");
@@ -134,7 +134,7 @@ client.on('message', async msg => {
         process.exit(1);
     }
     if (command === "shouldisleep" || command === "sleep") {
-        var currentHour = (new Date().getHours() + 8) % 24;
+        const currentHour = (new Date().getHours() + 8) % 24;
         console.log("current hour = " + currentHour);
         if (currentHour < 6 || currentHour > 21) {
             let preset = config.SLEEP_MESSAGES[Math.floor(Math.random() * config.SLEEP_MESSAGES.length)];
@@ -142,22 +142,19 @@ client.on('message', async msg => {
             embed.setDescription(preset.body.replace("${username}", msg.author.username));
             if (currentHour < 6 && currentHour > 2) embed.setThumbnail(config.SLEEP_LATE[Math.floor(Math.random() * config.SLEEP_LATE.length)]);
             else embed.setThumbnail(config.SLEEP_IMAGES[Math.floor(Math.random() * config.SLEEP_IMAGES.length)]);
-            msg.channel.send(embed);
+            await msg.channel.send(embed);
             console.log("told " + msg.author.username + " to to goto sleep");
         } else {
             let embed = new Discord.RichEmbed().setTitle("Get some **coffee**!").setColor(0x6f4e37);
             embed.setDescription("There is so much to do, better go get a cup of coffee **" + msg.author.username + "**");
             embed.setThumbnail("https://res.cloudinary.com/chatboxzy/image/upload/v1573747645/coffee.png");
-            msg.channel.send(embed);
+            await msg.channel.send(embed);
             console.log("told " + msg.author.username + " to not to goto sleep");
         }
     }
-    if (command === "clear") {
-        msg.channel.bulkDelete(parseInt(args[0]))
-            .then(messages => {
-                msg.reply("Deleted " + message.size() + " messages");
-            }).catch(console.error);
-    }
+    if (command === "clear") msg.channel.bulkDelete(parseInt(args[0])).then(messages => {
+        msg.reply("Deleted " + messages.keyArray().length + " messages");
+    }).catch(console.error);
     console.log("====== Message Processed, Elapsed time = " + (Date.now() - startTime) + "ms\n");
 });
 
@@ -174,38 +171,28 @@ client.on('error', error => {
     console.log("Error Occured " + error.message);
 })
 
-client.login(process.env.TOKEN);
+client.login(process.env.TOKEN).then(() => console.log("Logged In"));
 
 function chainUpStdOut() {
     const hook = new Discord.WebhookClient('670179846005063681', 'UckNPFVsz2nJ4bUAMtPMq_z0jFL0d66YNaq-C3OhhvXEDtid1hBj4tAOp5WVM9hFYYYn');
-    var util = require('util')
-    var events = require('events')
-
-    function hook_stdout(callback) {
-        var old_write = process.stdout.write
-        process.stdout.write = (function (write) {
-            return function (string, encoding, fd) {
-                write.apply(process.stdout, arguments)
-                callback(string, encoding, fd)
-            }
-        })(process.stdout.write)
-        return () => process.stdout.write = old_write;
-    }
-
     let str = [];
-    hook_stdout((s) => {
+    let callback = (s) => {
         str.push(s);
         if (s.endsWith('\n')) {
-            hook.send("`[" + new Date().toISOString() + "] " + str.join("") + "`");
+            hook.send("`[" + new Date().toISOString() + "] " + str.join("") + "`").catch(() => null);
             str = [];
         }
-    });
+    };
+    process.stdout.write = (write => function (string, encoding, fd) {
+        write.apply(process.stdout, arguments)
+        callback(string, encoding, fd)
+    })(process.stdout.write);
 }
 
-function join(parts, seperators, formatLength) {
+function join(parts, separators) {
     let str = "";
     for (let i = 0; i < parts.length; i++) {
-        str += parts[i] + seperators[i % seperators.length];
+        str += parts[i] + separators[i % separators.length];
     }
     if (str.length > 0) str = str.substring(0, str.length - 1);
     return str;
@@ -235,5 +222,5 @@ async function sendLongMessage(channel, msg, objects) {
 }
 
 function sendBubblePop(msg, args) {
-    msg.channel.send("||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||\n".repeat(args[0] && !isNaN(args[0]) ? Math.min(24, Math.max(1, parseInt(args[0]))) : 8) + `Have fun, ${msg.author.username}! (๑•̀ㅂ•́)و`);
+    return msg.channel.send("||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||||pop||\n".repeat(args[0] && !isNaN(args[0]) ? Math.min(24, Math.max(1, parseInt(args[0]))) : 8) + `Have fun, ${msg.author.username}! (๑•̀ㅂ•́)و`);
 }
